refactor: pull out pages styles into separate file
