+++ src/components/Views/GlobalNames/index.js
@@ -1,0 +1,1 @@
+export { default } from './GlobalNames';
