+++ src/components/Views/Trustlist/Trustlist.jsx
@@ -5,7 +5,7 @@ import Loader from 'components/common/NE_Loader';
 import { intlNum } from 'utils/converter';
 import CopyText from 'components/common/NE_CopyText/CopyText';
 import { useNetwork } from 'hooks/useNetwork/useNetwork';
-import ErrorCard from 'components/common/NE_ErrorCard/ErrorCard';
+import ErrorCard from 'components/common/NE_ErrorCard';
 import PageHeader from 'components/Header/PageHeader';
 import DynamicPagination from 'components/common/NE_Pagination';
 // import DynamicPagination from 'components/Table/DynamicPagination';
