+++ src/components/common/NE_CopyText/index.js
@@ -1,2 +1,1 @@
-export * from './CopyText';
 export { default } from './CopyText';
