+++ src/components/common/NE_ErrorCard/index.js
@@ -1,0 +1,1 @@
+export { default } from './ErrorCard';
