+++ src/components/common/NE_ScrollToTop/index.js
@@ -1,0 +1,2 @@
+export { default } from './ScrollToTop';
+export * from './ScrollToTop';
