+++ src/hooks/index.js
@@ -1,5 +1,9 @@
+export * from './useCookie/useCookie';
 export * from './useDarkMode/useDarkMode';
 export * from './useEventListener/useEventListener';
-export * from './useMediaQuery/useMediaQuery';
-export * from './useStorage/useStorage';
 export * from './useIsMobile/useIsMobile';
+export * from './useIsTouchDevice/useIsTouchDevice';
+export * from './useMediaQuery/useMediaQuery';
+export * from './useNetwork/useNetwork';
+export * from './useStorage/useStorage';
+export * from './useWindowSize/useWindowSize';
