+++ src/pages/_error.js
@@ -1,4 +1,4 @@
-import ErrorScreen from './error';
+import ErrorScreen from 'components/Views/error';
 
 ErrorScreen.getInitialProps = ({ res, err }) => {
   const statusCode = res ? res.statusCode : err ? err.statusCode : 404;
