+++ src/pages/blocks/index.js
@@ -1,0 +1,11 @@
+import PageHeader from 'components/Header/PageHeader';
+import BlocksView from 'components/Views/Blocks';
+
+export default function Blocks() {
+  return (
+    <>
+      <PageHeader page="blocks" />
+      <BlocksView />
+    </>
+  );
+}
