+++ src/pages/tokens/index.js
@@ -1,0 +1,11 @@
+import PageHeader from 'components/Header/PageHeader';
+import TokensView from 'components/Views/Tokens';
+
+export default function Tokens() {
+  return (
+    <>
+      <PageHeader page="Tokens" />
+      <TokensView />
+    </>
+  );
+}
