+++ src/pages/transactions/index.js
@@ -1,0 +1,11 @@
+import PageHeader from 'components/Header/PageHeader';
+import TransactionsView from 'components/Views/Transactions';
+
+export default function Transactions() {
+  return (
+    <>
+      <PageHeader page={'transactions'} />
+      <TransactionsView />
+    </>
+  );
+}
